requirements.txt
# FastAPI & Web Framework
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
python-multipart==0.0.9

# Data Validation & Models
pydantic==2.9.2

# HTTP Requests & Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
fake-useragent==1.5.1

# LLM & Structured Extraction (Lab 5)
openai>=1.52.0
instructor>=1.6.0
python-dotenv==1.0.1

# Vector Database (Lab 4 - Pinecone)
pinecone>=5.0.0
tiktoken>=0.7.0

# Optional: Browser Automation (Lab 1 - if re-scraping needed)
# selenium==4.25.0
# playwright==1.48.0
# tenacity==9.0.0

# Optional: Airflow (Lab 2 & 3)
# apache-airflow==2.10.0